Cross-temporal Reconciliation with Machine Learning

This function performs machine-learning–based cross-temporal forecast reconciliation for linearly constrained multiple time series (Rombouts et al., 2024). Reconciled forecasts are obtained by fitting non-linear models that map base forecasts across both temporal and cross-sectional dimensions to bottom-level high-frequency series. Fully coherent forecasts across all temporal and cross-sectional linear combinations are then derived by cross-temporal bottom-up. While the approach is designed for hierarchical and grouped structures, in the case of general linearly constrained time series it can be applied within the broader reconciliation framework described by Girolimetto and Di Fonzo (2024).

Usage

# Reconciled forecasts
ctrml(base, hat, obs, agg_mat, agg_order, tew = "sum", features = "all",
      approach = "randomForest", params = NULL, tuning = NULL,
      fit = NULL, sntz = FALSE, round = FALSE)

# Pre-trained reconciled ML models
ctrml_fit(hat, obs, agg_mat, agg_order, tew = "sum", features = "all",
          approach = "randomForest", params = NULL, tuning = NULL)

Arguments

base

A (\(n \times h(k^\ast+m)\)) numeric matrix containing the base forecasts to be reconciled; \(n\) is the total number of variables, \(m\) is the maximum aggregation order, and \(k^\ast\) is the sum of a chosen subset of the \(p - 1\) factors of \(m\) (excluding \(m\) itself), and \(h\) is the forecast horizon for the lowest frequency time series. The row identifies a time series, and the forecasts in each row are ordered from the lowest frequency (most temporally aggregated) to the highest frequency.

hat

A (\(n \times N(k^\ast+m)\)) numeric matrix containing the base forecasts ordered from lowest to highest frequency; \(N\) is the training length for the lowest frequency time series. The row identifies a time series, and the forecasts in each row are ordered from the lowest frequency (most temporally aggregated) to the highest frequency. These forecasts are used to train the ML approach.

obs

A (\(n_b \times Nm\)) numeric matrix containing (observed) values for the highest frequency series (\(k = 1\)); \(n_b\) is the total number of high-frequency bottom variables. These values are used to train the ML approach.

agg_mat

A (\(n_a \times n_b\)) numeric matrix representing the cross-sectional aggregation matrix. It maps the \(n_b\) bottom-level (free) variables into the \(n_a\) upper (constrained) variables.

agg_order

Highest available sampling frequency per seasonal cycle (max. order of temporal aggregation, \(m\)), or a vector representing a subset of \(p\) factors of \(m\).

tew

A string specifying the type of temporal aggregation. Options include: "sum" (simple summation, default), "avg" (average), "first" (first value of the period), and "last" (last value of the period).

features

Character string specifying which features are used for model training. Options include "all" (see Rombouts et al. 2025), and "compact" (see Rombouts et al. 2025, default).

approach

Character string specifying the machine learning method used for reconciliation. Options are:

"randomForest" (default): Random Forest algorithm (see the randomForest package).

"xgboost": Extreme Gradient Boosting (see the xgboost package).

"lightgbm": Light Gradient Boosting Machine (see the lightgbm package).

"mlr3": Any regression learner available in the mlr3 package. The learner must be specified via params, e.g. params = list(.key = "regr.ranger").

params

Optional list of additional parameters passed to the chosen ML approach These may include algorithm-specific hyperparameters for randomForest, xgboost, lightgbm, or learner options for mlr3. When approach = "mlr3", the list must include .key to select the learner (e.g. .key = "regr.ranger", default).

tuning

Optional list specifying tuning options when using the mlr3tuning::mlr3tuning framework (e.g., terminators, search spaces). The argument format follows mlr3tuning::auto_tuner, except that the learner is set through params.

fit

A pre-trained ML reconciliation model (see, extract_reconciled_ml). If supplied, training data (hat, obs) are not required.

sntz

Logical. If TRUE, enforces non-negativity on reconciled forecasts using the heuristic "set-negative-to-zero" (Di Fonzo and Girolimetto, 2023). Default is FALSE.

round

Logical. If TRUE, reconciled forecasts are rounded to integer values and coherence is ensured via a bottom-up adjustment. Default is FALSE.

Value

ctrml returns a cross-temporal reconciled forecast matrix with the same dimensions, along with attributes containing the fitted model and reconciliation settings (see, FoReco::recoinfo and extract_reconciled_ml).

ctrml_fit returns a fitted object that can be reused for reconciliation on new base forecasts.

References

Di Fonzo, T. and Girolimetto, D. (2023), Spatio-temporal reconciliation of solar forecasts, Solar Energy, 251, 13–29. doi:10.1016/j.solener.2023.01.003

Girolimetto, D. and Di Fonzo, T. (2023), Point and probabilistic forecast reconciliation for general linearly constrained multiple time series, Statistical Methods & Applications, 33, 581-607. doi:10.1007/s10260-023-00738-6 .

Rombouts, J., Ternes, M., and Wilms, I. (2025). Cross-temporal forecast reconciliation at digital platforms with machine learning. International Journal of Forecasting, 41(1), 321-344. doi:10.1016/j.ijforecast.2024.05.008

Examples

# m: quarterly temporal aggregation order
m <- 4
te_set <- tetools(m)$set

# agg_mat: simple aggregation matrix, A = B + C
agg_mat <- t(c(1,1))
dimnames(agg_mat) <- list("A", c("B", "C"))

# te_fh: minimum forecast horizon per temporal aggregate
te_fh <- m/te_set

# N_hat: dimension for the lowest-frequency (k = m) training set
N_hat <- 16

# bts_mean: mean for the Normal draws used to simulate data
bts_mean <- 5

# hat: a training (base forecasts) feautures matrix
hat <- rbind(
  rnorm(sum(te_fh)*N_hat, rep(2*te_set*bts_mean, N_hat*te_fh)),  # Series A
  rnorm(sum(te_fh)*N_hat, rep(te_set*bts_mean,  N_hat*te_fh)),   # Series B
  rnorm(sum(te_fh)*N_hat, rep(te_set*bts_mean,  N_hat*te_fh))    # Series C
)
rownames(hat) <- c("A", "B", "C")

# obs: (observed) values for the highest-frequency bottom-level series
# (B and C with k = 1)
obs <- rbind(
  rnorm(m*N_hat, bts_mean),  # Observed for series B
  rnorm(m*N_hat, bts_mean)   # Observed for series C
)
rownames(obs) <- c("B", "C")

# h: base forecast horizon at the lowest-frequency series (k = m)
h <- 2

# base: base forecasts matrix
base <- rbind(
  rnorm(sum(te_fh)*h, rep(2*te_set*bts_mean, h*te_fh)),  # Base for A
  rnorm(sum(te_fh)*h, rep(te_set*bts_mean,  h*te_fh)),   # Base for B
  rnorm(sum(te_fh)*h, rep(te_set*bts_mean,  h*te_fh))    # Base for C
)
rownames(base) <- c("A", "B", "C")

##########################################################################
# Different ML approaches
##########################################################################
# XGBoost Reconciliation (xgboost pkg)
reco <- ctrml(base = base, hat = hat, obs = obs, agg_order = m,
              agg_mat = agg_mat, approach = "xgboost")

# XGBoost Reconciliation with Tweedie loss function (xgboost pkg)
reco <- ctrml(base = base, hat = hat, obs = obs, agg_order = m,
              agg_mat = agg_mat, approach = "xgboost",
              params =  list(
                eta = 0.3, colsample_bytree = 1, min_child_weight = 1,
                max_depth = 6, gamma = 0, subsample = 1,
                objective = "reg:tweedie", # Tweedie regression objective
                tweedie_variance_power = 1.5 # Tweedie power parameter
              ))

# LightGBM Reconciliation (lightgbm pkg)
reco <- ctrml(base = base, hat = hat, obs = obs, agg_order = m,
              agg_mat = agg_mat, approach = "lightgbm")

# Random Forest Reconciliation (randomForest pkg)
reco <- ctrml(base = base, hat = hat, obs = obs, agg_order = m,
              agg_mat = agg_mat, approach = "randomForest")

# Using the mlr3 pkg:
# With 'params = list(.key = mlr_learners)' we can specify different
# mlr_learners implemented in mlr3 such as "regr.ranger" for Random Forest,
# "regr.xgboost" for XGBoost, and others.
reco <- ctrml(base = base, hat = hat, obs = obs, agg_order = m,
              agg_mat = agg_mat, approach = "mlr3",
              # choose mlr3 learner (here Random Forest via ranger)
              params = list(.key = "regr.ranger"))
# \donttest{
# With mlr3 we can also tune our parameters: e.g. explore mtry in [1,4].
# We can reduce excessive logging by calling:
# if(requireNamespace("lgr", quietly = TRUE)){
#   lgr::get_logger("mlr3")$set_threshold("warn")
#   lgr::get_logger("bbotk")$set_threshold("warn")
# }
reco <- ctrml(base = base, hat = hat, obs = obs, agg_order = m,
              agg_mat = agg_mat, approach = "mlr3",
              params = list(
                .key = "regr.ranger",
                # number of features tried at each split
                mtry = paradox::to_tune(paradox::p_int(1, 4))
              ),
              tuning = list(
                # stop after 10 evaluations
                terminator = mlr3tuning::trm("evals", n_evals = 10)
              ))
#> Warning: row names were found from a short variable and have been discarded
#> INFO  [17:13:30.740] [bbotk] Starting to optimize 1 parameter(s) with '<OptimizerBatchRandomSearch>' and '<TerminatorEvals> [n_evals=10, k=0]'
#> INFO  [17:13:30.755] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:30.760] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:30.786] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:30.825] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:30.884] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:30.922] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:30.962] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:31.000] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:31.041] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:31.084] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:31.125] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:31.168] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:31.213] [mlr3] Finished benchmark
#> INFO  [17:13:31.239] [bbotk] Result of batch 1:
#> INFO  [17:13:31.241] [bbotk]  mtry   regr.mse warnings errors runtime_learners
#> INFO  [17:13:31.241] [bbotk]     3 0.04857838        0      0            0.165
#> INFO  [17:13:31.241] [bbotk]     4 0.04362256        0      0            0.186
#> INFO  [17:13:31.241] [bbotk]                                 uhash
#> INFO  [17:13:31.241] [bbotk]  4e7be88e-8c0a-42dd-864c-526c15022d79
#> INFO  [17:13:31.241] [bbotk]  d2ce5b96-4bca-4ad7-a8b9-841bece255ea
#> INFO  [17:13:31.243] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:31.246] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:31.253] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:31.295] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:31.335] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:31.376] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:31.417] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:31.459] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:31.502] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:31.544] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:31.583] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:31.621] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:31.660] [mlr3] Finished benchmark
#> INFO  [17:13:31.676] [bbotk] Result of batch 2:
#> INFO  [17:13:31.677] [bbotk]  mtry   regr.mse warnings errors runtime_learners
#> INFO  [17:13:31.677] [bbotk]     4 0.04312236        0      0            0.179
#> INFO  [17:13:31.677] [bbotk]     3 0.05046774        0      0            0.176
#> INFO  [17:13:31.677] [bbotk]                                 uhash
#> INFO  [17:13:31.677] [bbotk]  69ab42a4-84fc-4dfc-b04b-e783e6311da7
#> INFO  [17:13:31.677] [bbotk]  c2ef9251-4b00-4e95-aaa5-cfa29a62f78a
#> INFO  [17:13:31.679] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:31.681] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:31.685] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:31.727] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:31.769] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:31.807] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:31.847] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:31.887] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:31.926] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:31.965] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:32.004] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:32.053] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:32.093] [mlr3] Finished benchmark
#> INFO  [17:13:32.109] [bbotk] Result of batch 3:
#> INFO  [17:13:32.110] [bbotk]  mtry   regr.mse warnings errors runtime_learners
#> INFO  [17:13:32.110] [bbotk]     3 0.05180063        0      0            0.169
#> INFO  [17:13:32.110] [bbotk]     3 0.04817492        0      0            0.166
#> INFO  [17:13:32.110] [bbotk]                                 uhash
#> INFO  [17:13:32.110] [bbotk]  fb4c9923-2d79-4ec4-967d-13d49b87b402
#> INFO  [17:13:32.110] [bbotk]  0db57325-398c-45ed-a3c1-20c25b683a08
#> INFO  [17:13:32.113] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:32.115] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:32.118] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:32.160] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:32.201] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:32.242] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:32.288] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:32.336] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:32.377] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:32.417] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:32.459] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:32.500] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:32.544] [mlr3] Finished benchmark
#> INFO  [17:13:32.687] [bbotk] Result of batch 4:
#> INFO  [17:13:32.688] [bbotk]  mtry   regr.mse warnings errors runtime_learners
#> INFO  [17:13:32.688] [bbotk]     4 0.04579038        0      0            0.189
#> INFO  [17:13:32.688] [bbotk]     4 0.04480842        0      0            0.182
#> INFO  [17:13:32.688] [bbotk]                                 uhash
#> INFO  [17:13:32.688] [bbotk]  a72297f4-cc2d-41ba-8432-b1754568bb7d
#> INFO  [17:13:32.688] [bbotk]  72da39d9-6b3d-4d18-9283-ff48a3426f74
#> INFO  [17:13:32.691] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:32.693] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:32.697] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:32.738] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:32.779] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:32.821] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:32.862] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:32.906] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:32.938] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:32.968] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:32.997] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:33.026] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:33.056] [mlr3] Finished benchmark
#> INFO  [17:13:33.073] [bbotk] Result of batch 5:
#> INFO  [17:13:33.074] [bbotk]  mtry  regr.mse warnings errors runtime_learners
#> INFO  [17:13:33.074] [bbotk]     4 0.0444781        0      0            0.182
#> INFO  [17:13:33.074] [bbotk]     1 0.1160210        0      0            0.121
#> INFO  [17:13:33.074] [bbotk]                                 uhash
#> INFO  [17:13:33.074] [bbotk]  e96a40eb-4ccf-48ac-8f9e-e4576c7fd4b8
#> INFO  [17:13:33.074] [bbotk]  92200bbc-e5dd-4938-9499-134bcb0658d5
#> INFO  [17:13:33.082] [bbotk] Finished optimizing after 10 evaluation(s)
#> INFO  [17:13:33.082] [bbotk] Result:
#> INFO  [17:13:33.083] [bbotk]   mtry learner_param_vals  x_domain   regr.mse
#> INFO  [17:13:33.083] [bbotk]  <int>             <list>    <list>      <num>
#> INFO  [17:13:33.083] [bbotk]      4          <list[2]> <list[1]> 0.04312236
#> Warning: row names were found from a short variable and have been discarded
#> INFO  [17:13:33.200] [bbotk] Starting to optimize 1 parameter(s) with '<OptimizerBatchRandomSearch>' and '<TerminatorEvals> [n_evals=10, k=0]'
#> INFO  [17:13:33.205] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:33.207] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:33.210] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:33.239] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:33.268] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:33.296] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:33.325] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:33.355] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:33.397] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:33.566] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:33.610] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:33.656] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:33.699] [mlr3] Finished benchmark
#> INFO  [17:13:33.714] [bbotk] Result of batch 1:
#> INFO  [17:13:33.715] [bbotk]  mtry   regr.mse warnings errors runtime_learners
#> INFO  [17:13:33.715] [bbotk]     1 0.14308281        0      0            0.119
#> INFO  [17:13:33.715] [bbotk]     4 0.04560309        0      0            0.315
#> INFO  [17:13:33.715] [bbotk]                                 uhash
#> INFO  [17:13:33.715] [bbotk]  b0b3179e-a555-438d-a831-3dbdac4f1565
#> INFO  [17:13:33.715] [bbotk]  207617f7-494e-4a14-a226-1b7ee37dd2e8
#> INFO  [17:13:33.718] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:33.720] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:33.723] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:33.758] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:33.794] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:33.828] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:33.862] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:33.898] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:33.936] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:33.977] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:34.016] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:34.056] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:34.094] [mlr3] Finished benchmark
#> INFO  [17:13:34.111] [bbotk] Result of batch 2:
#> INFO  [17:13:34.113] [bbotk]  mtry   regr.mse warnings errors runtime_learners
#> INFO  [17:13:34.113] [bbotk]     2 0.07159014        0      0            0.149
#> INFO  [17:13:34.113] [bbotk]     3 0.05196688        0      0            0.167
#> INFO  [17:13:34.113] [bbotk]                                 uhash
#> INFO  [17:13:34.113] [bbotk]  f1205d02-8c7c-4421-be13-fba4d033c717
#> INFO  [17:13:34.113] [bbotk]  f257a5cf-e9d3-45e8-bff7-ccd2223598e1
#> INFO  [17:13:34.115] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:34.118] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:34.122] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:34.150] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:34.178] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:34.207] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:34.245] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:34.275] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:34.304] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:34.332] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:34.360] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:34.388] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:34.417] [mlr3] Finished benchmark
#> INFO  [17:13:34.433] [bbotk] Result of batch 3:
#> INFO  [17:13:34.434] [bbotk]  mtry  regr.mse warnings errors runtime_learners
#> INFO  [17:13:34.434] [bbotk]     1 0.1421412        0      0            0.128
#> INFO  [17:13:34.434] [bbotk]     1 0.1440664        0      0            0.116
#> INFO  [17:13:34.434] [bbotk]                                 uhash
#> INFO  [17:13:34.434] [bbotk]  a7b97a11-8baa-4db6-a3e9-6cc7b0181f69
#> INFO  [17:13:34.434] [bbotk]  96977252-5886-40c5-a124-7552d53c116b
#> INFO  [17:13:34.436] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:34.439] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:34.442] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:34.470] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:34.499] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:34.528] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:34.557] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:34.586] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:34.617] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:34.645] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:34.674] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:34.701] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:34.730] [mlr3] Finished benchmark
#> INFO  [17:13:34.753] [bbotk] Result of batch 4:
#> INFO  [17:13:34.754] [bbotk]  mtry  regr.mse warnings errors runtime_learners
#> INFO  [17:13:34.754] [bbotk]     1 0.1446619        0      0            0.122
#> INFO  [17:13:34.754] [bbotk]     1 0.1423207        0      0            0.119
#> INFO  [17:13:34.754] [bbotk]                                 uhash
#> INFO  [17:13:34.754] [bbotk]  5229e114-8d74-4537-a490-915fcc70dc0b
#> INFO  [17:13:34.754] [bbotk]  a4871d11-13b4-471b-a2c3-055f0eab7ee4
#> INFO  [17:13:34.757] [bbotk] Evaluating 2 configuration(s)
#> INFO  [17:13:34.759] [mlr3] Running benchmark with 10 resampling iterations
#> INFO  [17:13:34.763] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:34.793] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:34.823] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:34.850] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:34.879] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:34.907] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 1/5)
#> INFO  [17:13:34.941] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 2/5)
#> INFO  [17:13:34.976] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 3/5)
#> INFO  [17:13:35.009] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 4/5)
#> INFO  [17:13:35.044] [mlr3] Applying learner 'regr.ranger' on task 'tsk_i' (iter 5/5)
#> INFO  [17:13:35.078] [mlr3] Finished benchmark
#> INFO  [17:13:35.094] [bbotk] Result of batch 5:
#> INFO  [17:13:35.095] [bbotk]  mtry   regr.mse warnings errors runtime_learners
#> INFO  [17:13:35.095] [bbotk]     1 0.14752386        0      0            0.121
#> INFO  [17:13:35.095] [bbotk]     2 0.07165827        0      0            0.145
#> INFO  [17:13:35.095] [bbotk]                                 uhash
#> INFO  [17:13:35.095] [bbotk]  d508f899-2c94-4daa-b2ed-98a16c162472
#> INFO  [17:13:35.095] [bbotk]  e943feb0-70cc-4bec-b69f-aed8e1ab9891
#> INFO  [17:13:35.102] [bbotk] Finished optimizing after 10 evaluation(s)
#> INFO  [17:13:35.103] [bbotk] Result:
#> INFO  [17:13:35.104] [bbotk]   mtry learner_param_vals  x_domain   regr.mse
#> INFO  [17:13:35.104] [bbotk]  <int>             <list>    <list>      <num>
#> INFO  [17:13:35.104] [bbotk]      4          <list[2]> <list[1]> 0.04560309
# }
##########################################################################
# Usage with pre-trained models
##########################################################################
# Pre-trained machine learning models (e.g., omit the base param)
mdl <- ctrml_fit(hat = hat, obs = obs, agg_order = m, agg_mat = agg_mat,
                 approach = "xgboost")

# Pre-trained machine learning models with base param
reco <- ctrml(base = base, hat = hat, obs = obs, agg_order = m,
              agg_mat = agg_mat, approach = "xgboost")
mdl2 <- extract_reconciled_ml(reco)

# New base forecasts matrix
base_new <- rbind(
  rnorm(sum(te_fh)*h, rep(2*te_set*bts_mean, h*te_fh)),  # Base for A
  rnorm(sum(te_fh)*h, rep(te_set*bts_mean,  h*te_fh)),   # Base for B
  rnorm(sum(te_fh)*h, rep(te_set*bts_mean,  h*te_fh))    # Base for C
)
reco_new <- ctrml(base = base_new, fit = mdl, agg_order = m,
                  agg_mat = agg_mat)